+++ orix/data/_registry.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018-2022 the orix developers
+#
+# This file is part of orix.
+#
+# orix is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# orix is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with orix.  If not, see <http://www.gnu.org/licenses/>.
+
+# fmt: off
+registry_hashes = {
+    "sdss/sdss_austenite.h5": "a650c4b49cafd7019644952dbc8cbb85741119e0d69edbf8b030d50ad0fa855c",
+    "sdss/sdss_ferrite_austenite.ang": "ac2fc2fc2b8c9c598703c96f9888677bccc9cd88cddd5e78735653fd5504e371",
+    "ti_orientations/ti_orientations.ctf": "d8228c58af45ed70f3617b8865c13878bb1fc40b3d60163042be4b761105dcde",
+}
+registry_urls = {
+    "sdss/sdss_austenite.h5": "https://github.com/pyxem/orix-data/raw/main/sdss_austenite/sdss_austenite.h5",
+    "sdss/sdss_ferrite_austenite.ang": "https://github.com/pyxem/orix-data/raw/main/sdss_ferrite_austenite/sdss_ferrite_austenite.ang",
+    "ti_orientations/ti_orientations.ctf": "https://github.com/pyxem/orix-data/raw/main/ti_orientations/Ti_orientations.ctf",
+}
+# fmt: on