+++ orix/quaternion/quaternion.py
@@ -19,6 +19,7 @@
 import warnings
 
 import dask.array as da
+from dask.diagnostics import ProgressBar
 import numpy as np
 import quaternion
 
@@ -58,11 +59,11 @@ class Quaternion(Object3d):
     :math:`v' = q * v * q^{-1}` and follows as:
 
     .. math::
-       v'_x = x(a^2 + b^2 - c^2 - d^2) + 2z(a * c + b * d) + y(b * c - a * d)
+       v'_x = x(a^2 + b^2 - c^2 - d^2) + 2(z(a * c + b * d) + y(b * c - a * d))
 
-       v'_y = y(a^2 - b^2 + c^2 - d^2) + 2x(a * d + b * c) + z(c * d - a * b)
+       v'_y = y(a^2 - b^2 + c^2 - d^2) + 2(x(a * d + b * c) + z(c * d - a * b))
 
-       v'_z = z(a^2 - b^2 - c^2 + d^2) + 2y(a * b + c * d) + x(b * d - a * c)
+       v'_z = z(a^2 - b^2 - c^2 + d^2) + 2(y(a * b + c * d) + x(b * d - a * c))
 
     Attributes
     ----------
@@ -222,13 +223,24 @@ class Quaternion(Object3d):
         w_max = np.argmax(w)
         return self.__class__(v[:, w_max])
 
-    def outer(self, other):
+    def outer(self, other, lazy=False, chunk_size=20, progressbar=True):
         """Compute the outer product of this quaternion and the other
         quaternion or vector.
 
         Parameters
         ----------
         other : orix.quaternion.Quaternion or orix.vector.Vector3d
+        lazy : bool, optional
+            Whether to computer this computation using Dask. This option
+            can be used to reduce memory usage when working with large
+            arrays. Default is False.
+        chunk_size : int, optional
+            When using `lazy` computation, `chunk_size` represents the
+            number of objects per axis for each input to include in each
+            iteration of the computation. Default is 20.
+        progressbar : bool, optional
+            Whether to show a progressbar during computation if `lazy`
+            is True. Default is True.
 
         Returns
         -------
@@ -236,20 +248,39 @@ class Quaternion(Object3d):
         """
 
         if isinstance(other, Quaternion):
-            q1 = quaternion.from_float_array(self.data)
-            q2 = quaternion.from_float_array(other.data)
-            # np.outer works with flattened array
-            q = np.outer(q1, q2).reshape(q1.shape + q2.shape)
-            return other.__class__(quaternion.as_float_array(q))
+            if lazy:
+                darr = self._outer_dask(other, chunk_size=chunk_size)
+                arr = np.empty(darr.shape)
+                if progressbar:
+                    with ProgressBar():
+                        da.store(darr, arr)
+                else:
+                    da.store(darr, arr)
+            else:
+                q1 = quaternion.from_float_array(self.data)
+                q2 = quaternion.from_float_array(other.data)
+                # np.outer works with flattened array
+                q = np.outer(q1, q2).reshape(q1.shape + q2.shape)
+                arr = quaternion.as_float_array(q)
+            return other.__class__(arr)
         elif isinstance(other, Vector3d):
-            q = quaternion.from_float_array(self.data)
-            v = quaternion.rotate_vectors(q, other.data)
+            if lazy:
+                darr = self._outer_dask(other, chunk_size=chunk_size)
+                arr = np.empty(darr.shape)
+                if progressbar:
+                    with ProgressBar():
+                        da.store(darr, arr)
+                else:
+                    da.store(darr, arr)
+            else:
+                q = quaternion.from_float_array(self.data)
+                arr = quaternion.rotate_vectors(q, other.data)
             if isinstance(other, Miller):
-                m = other.__class__(xyz=v, phase=other.phase)
+                m = other.__class__(xyz=arr, phase=other.phase)
                 m.coordinate_format = other.coordinate_format
                 return m
             else:
-                return other.__class__(v)
+                return other.__class__(arr)
         else:
             raise NotImplementedError(
                 "This operation is currently not avaliable in orix, please use outer "
@@ -258,17 +289,18 @@ class Quaternion(Object3d):
 
     def _outer_dask(self, other, chunk_size=20):
         """Compute the product of every quaternion in this instance to
-        every quaternion in another instance, returned as a Dask array.
+        every quaternion or vector in another instance, returned as a
+        Dask array.
 
-        This is also known as the Hamilton product.
+        For quaternion-quaternion multiplication, this is known as the
+        Hamilton product.
 
         Parameters
         ----------
-        other : orix.quaternion.Quaternion
+        other : orix.quaternion.Quaternion or orix.vector.Vector3d
         chunk_size : int, optional
-            Number of quaternions per axis in each quaternion instance
-            to include in each iteration of the computation. Default is
-            20.
+            Number of objects per axis for each input to include in each
+            iteration of the computation. Default is 20.
 
         Returns
         -------
@@ -276,9 +308,15 @@ class Quaternion(Object3d):
 
         Notes
         -----
-        To get a new quaternion from the returned array `qarr`, do
-        `q = Quaternion(qarr.compute())`.
+        For quaternion-quaternion multiplication, to create a new
+        quaternion from the returned array `arr`, do
+        `q = Quaternion(arr.compute())`. Likewise for quaternion-vector
+        multiplication, to create a new vector from the returned array
+        do `v = Vector3d(arr.compute())`.
         """
+        if not isinstance(other, (Quaternion, Vector3d)):
+            raise TypeError("Other must be Quaternion or Vector3d.")
+
         ndim1 = len(self.shape)
         ndim2 = len(other.shape)
 
@@ -286,49 +324,72 @@ class Quaternion(Object3d):
         chunks1 = (chunk_size,) * ndim1 + (-1,)
         chunks2 = (chunk_size,) * ndim2 + (-1,)
 
-        # Get quaternion parameters as dask arrays to be computed later
-        q1 = da.from_array(self.data, chunks=chunks1)
-        a1, b1, c1, d1 = q1[..., 0], q1[..., 1], q1[..., 2], q1[..., 3]
-        q2 = da.from_array(other.data, chunks=chunks2)
-        a2, b2, c2, d2 = q2[..., 0], q2[..., 1], q2[..., 2], q2[..., 3]
-
         # Dask has no dask.multiply.outer(), use dask.array.einsum
         # Summation subscripts
         str1 = "abcdefghijklm"[:ndim1]  # Max. object dimension of 13
         str2 = "nopqrstuvwxyz"[:ndim2]
         sum_over = f"...{str1},{str2}...->{str1 + str2}"
 
+        # Get quaternion parameters as dask arrays to be computed later
+        q1 = da.from_array(self.data, chunks=chunks1)
+        a1, b1, c1, d1 = q1[..., 0], q1[..., 1], q1[..., 2], q1[..., 3]
+
         # We silence dask's einsum performance warnings for "small"
         # chunk sizes, since using the chunk sizes suggested floods
         # memory
         warnings.filterwarnings("ignore", category=da.PerformanceWarning)
 
-        # fmt: off
-        a = (
-            + da.einsum(sum_over, a1, a2)
-            - da.einsum(sum_over, b1, b2)
-            - da.einsum(sum_over, c1, c2)
-            - da.einsum(sum_over, d1, d2)
-        )
-        b = (
-            + da.einsum(sum_over, b1, a2)
-            + da.einsum(sum_over, a1, b2)
-            - da.einsum(sum_over, d1, c2)
-            + da.einsum(sum_over, c1, d2)
-        )
-        c = (
-            + da.einsum(sum_over, c1, a2)
-            + da.einsum(sum_over, d1, b2)
-            + da.einsum(sum_over, a1, c2)
-            - da.einsum(sum_over, b1, d2)
-        )
-        d = (
-            + da.einsum(sum_over, d1, a2)
-            - da.einsum(sum_over, c1, b2)
-            + da.einsum(sum_over, b1, c2)
-            + da.einsum(sum_over, a1, d2)
-        )
-        # fmt: on
+        if isinstance(other, Quaternion):
+            q2 = da.from_array(other.data, chunks=chunks2)
+            a2, b2, c2, d2 = q2[..., 0], q2[..., 1], q2[..., 2], q2[..., 3]
+            # fmt: off
+            a = (
+                + da.einsum(sum_over, a1, a2)
+                - da.einsum(sum_over, b1, b2)
+                - da.einsum(sum_over, c1, c2)
+                - da.einsum(sum_over, d1, d2)
+            )
+            b = (
+                + da.einsum(sum_over, b1, a2)
+                + da.einsum(sum_over, a1, b2)
+                - da.einsum(sum_over, d1, c2)
+                + da.einsum(sum_over, c1, d2)
+            )
+            c = (
+                + da.einsum(sum_over, c1, a2)
+                + da.einsum(sum_over, d1, b2)
+                + da.einsum(sum_over, a1, c2)
+                - da.einsum(sum_over, b1, d2)
+            )
+            d = (
+                + da.einsum(sum_over, d1, a2)
+                - da.einsum(sum_over, c1, b2)
+                + da.einsum(sum_over, b1, c2)
+                + da.einsum(sum_over, a1, d2)
+            )
+            # fmt: on
+            out = da.stack((a, b, c, d), axis=-1)
+        else:  # Vector3d
+            v2 = da.from_array(other.data, chunks=chunks2)
+            x2, y2, z2 = v2[..., 0], v2[..., 1], v2[..., 2]
+            # fmt: off
+            x = (
+                + da.einsum(sum_over, a1**2 + b1**2 - c1**2 - d1**2, x2)
+                + da.einsum(sum_over, a1 * c1 + b1 * d1, z2) * 2
+                + da.einsum(sum_over, b1 * c1 - a1 * d1, y2) * 2
+            )
+            y = (
+                + da.einsum(sum_over, a1**2 - b1**2 + c1**2 - d1**2, y2)
+                + da.einsum(sum_over, a1 * d1 + b1 * c1, x2) * 2
+                + da.einsum(sum_over, c1 * d1 - a1 * b1, z2) * 2
+            )
+            z = (
+                + da.einsum(sum_over, a1**2 - b1**2 - c1**2 + d1**2, z2)
+                + da.einsum(sum_over, a1 * b1 + c1 * d1, y2) * 2
+                + da.einsum(sum_over, b1 * d1 - a1 * c1, x2) * 2
+            )
+            # fmt: on
+            out = da.stack((x, y, z), axis=-1)
 
         new_chunks = tuple(chunks1[:-1]) + tuple(chunks2[:-1]) + (-1,)
-        return da.stack((a, b, c, d), axis=-1).rechunk(new_chunks)
+        return out.rechunk(new_chunks)